fix(l4): cast gradient ratio to float to satisfy mypy (no behavior change)
The inf-on-zero/NaN branch assigns float("inf") while the division branch
yielded a numpy floating, which mypy flagged as an incompatible assignment.
Wrapping the division in float() makes both branches Python float; the recorded
value is identical. Diagnostic-only path; baselines unaffected.

+++ xpcsjax/optimization/nlsq/gradient_monitor.py
@@ -308,7 +308,7 @@ class GradientCollapseMonitor:
         # collapse rather than masking it with a tiny epsilon.
         denom = per_angle_grad_norm
         if denom > 0 and np.isfinite(denom):
-            ratio = physical_grad_norm / denom
+            ratio = float(physical_grad_norm / denom)
         else:
             ratio = float("inf")
